feat: Redesign gpa calculator
Made it dynamic so modules were not hardcoded.
I need to look into a cleaner way to do this
