+++ src/app/calculators/gpa/page.tsx
@@ -1,327 +1,278 @@
 "use client"
 
-import { Label } from "@radix-ui/react-label"
-import React, { type Dispatch, type SetStateAction, useEffect, useState } from "react"
-import { Slider } from "~/components/ui/slider"
+import { useFieldArray, type UseFieldArrayReturn, useForm, type UseFormReturn } from "react-hook-form"
+import { z } from "zod"
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useEffect, useState } from "react"
+import { Form, FormField } from "~/components/ui/form"
+import { Card, CardContent, CardTitle } from "~/components/ui/card"
+import { Separator } from "~/components/ui/separator"
+import { Input } from "~/components/ui/input"
+import { cn, markToLabel, roundToDigits } from "~/lib/utils"
 import { Button } from "~/components/ui/button"
-import { useToast } from "~/hooks/use-toast"
+import { PlusIcon, TrashIcon } from "lucide-react"
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "~/components/ui/tooltip"
 
-type MarkType = {
-    semester5: {
-        computerVision: number,
-        researchProject: number,
-        softwareEngineering: number,
-        devops: number,
-        serverside: number,
-    },
-    semester6: {
-        legalIssues: number,
-        predictiveAnalytics: number,
-        projectDevelopment: number,
-        computerScience: number,
-        uxDesign: number,
-    }
-}
 
-const FIVE_CREDIT_PERCENTAGE = 5 / 30
+const FormSchema = z.object({
+    semesters: z.array(z.object({
+        name: z.string(),
+        subjects: z.array(z.object({
+            name: z.string(),
+            grade: z.number().min(0, "Grade must be at least 0").max(100, "Grade must be between 0 and 100"),
+            credits: z.number().min(1, "Credits must be greater than 0")
+        }))
+    })).min(1, "At least one semester is required")
+})
 
-const calculateS5Total = (markState: MarkType) => {
-    const marks = markState.semester5
-    let total = 0
+type FormType = z.infer<typeof FormSchema>
 
-    total += marks.computerVision * FIVE_CREDIT_PERCENTAGE
-    total += marks.researchProject * FIVE_CREDIT_PERCENTAGE * 2
-    total += marks.softwareEngineering * FIVE_CREDIT_PERCENTAGE
-    total += marks.devops * FIVE_CREDIT_PERCENTAGE
-    total += marks.serverside * FIVE_CREDIT_PERCENTAGE
+function Semester({
+    semesterIndex,
+    semestersFieldArray,
+    form,
+}: {
+    semesterIndex: number,
+    semestersFieldArray: UseFieldArrayReturn<FormType, "semesters">,
+    form: UseFormReturn<FormType>,
+}) {
+    const [gpa, setGpa] = useState(0)
 
-    // Rounded to 2 decimal places
-    total = Math.round(total * 100) / 100
-    return total
-}
+    const subjects = useFieldArray({
+        control: form.control,
+        name: `semesters.${semesterIndex}.subjects`
+    })
 
-const calculateS6Total = (markState: MarkType) => {
-    const marks = markState.semester6
-    let total = 0
+    // Calculate GPA on subjects change
+    useEffect(() => {
+        // Manually trigger validation for subjects
+        form.trigger(`semesters.${semesterIndex}.subjects`).then((isValid) => {
 
-    total += marks.legalIssues * FIVE_CREDIT_PERCENTAGE
-    total += marks.predictiveAnalytics * FIVE_CREDIT_PERCENTAGE
-    total += marks.projectDevelopment * FIVE_CREDIT_PERCENTAGE * 2
-    total += marks.computerScience * FIVE_CREDIT_PERCENTAGE
-    total += marks.uxDesign * FIVE_CREDIT_PERCENTAGE
+            // If subjects are not valid, set GPA to 0
+            if (!isValid) {
+                setGpa(0)
+                return
+            }
 
-    // Rounded to 2 decimal places
-    total = Math.round(total * 100) / 100
-    return total
-}
+            // Calculate GPA
+            const subjects = form.getValues(`semesters.${semesterIndex}.subjects`)
+            const totalCredits = subjects.reduce((acc, subject) => acc + subject.credits, 0)
+            const totalGrade = subjects.reduce((acc, subject) => acc + subject.grade * subject.credits / totalCredits, 0)
+            setGpa(roundToDigits(totalGrade, 2))
+        }).catch(() => {
+            // If any error occurs, set GPA to 0
+            setGpa(0)
+        })
+    }, [subjects, semesterIndex, form])
 
-const calculateTotal = (markState: MarkType) => {
-    const result = (calculateS5Total(markState) + calculateS6Total(markState)) / 2
-
-    // Rounded to 2 decimal places
-    return Math.round(result * 100) / 100
-}
-
-const getMarkLabel = (mark: number) => {
-    if (mark < 50) return "Class 3 Honours"
-    if (mark < 60) return "Class 2.2 Honours"
-    if (mark < 70) return "Class 2.1 Honours"
-    return "Class 1 Honours"
-}
-
-const Semester5Inputs: React.FC<{ markState: MarkType, setMarkState: Dispatch<SetStateAction<MarkType>> }> = ({ markState, setMarkState }) => {
     return (
-        <div className="flex flex-col mt-6 gap-4 w-full items-center">
-            <div className="grid w-full max-w-sm items-center gap-2">
-                <Label>Computer Vision</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester5.computerVision]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester5: {
-                                ...prev.semester5,
-                                computerVision: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester5.computerVision}%</p>
+        <Card key={semesterIndex} className="w-full border-emerald-400">
+            <CardTitle className="flex justify-center font-bold text-xl p-4">
+                <div className="flex-grow flex items-center">
+                    <Input
+                        {...form.register(`semesters.${semesterIndex}.name`)}
+                        className="w-fit"
+                        placeholder="Semester name"
+                    />
                 </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>Research in Computing</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester5.researchProject]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester5: {
-                                ...prev.semester5,
-                                researchProject: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester5.researchProject}%</p>
-                </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>Software Engineering</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester5.softwareEngineering]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester5: {
-                                ...prev.semester5,
-                                softwareEngineering: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester5.softwareEngineering}%</p>
-                </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>DevOps</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester5.devops]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester5: {
-                                ...prev.semester5,
-                                devops: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester5.devops}%</p>
-                </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>Server Side Scripting</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester5.serverside]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester5: {
-                                ...prev.semester5,
-                                serverside: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester5.serverside}%</p>
-                </div>
-            </div>
-        </div>
-    )
-}
+                <Tooltip>
+                    <TooltipTrigger asChild>
+                        <Button
+                            size="icon"
 
-const Semester6Inputs: React.FC<{ markState: MarkType, setMarkState: Dispatch<SetStateAction<MarkType>> }> = ({ markState, setMarkState }) => {
-    return (
-        <div className="flex flex-col mt-6 gap-4 w-full items-center">
-            <div className="grid w-full max-w-sm items-center gap-2">
-                <Label>Legal Issues in Computing</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester6.legalIssues]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester6: {
-                                ...prev.semester6,
-                                legalIssues: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester6.legalIssues}%</p>
+                            variant="destructive"
+                            onClick={() => semestersFieldArray.remove(semesterIndex)}
+                        >
+                            <TrashIcon />
+                        </Button>
+                    </TooltipTrigger>
+                    <TooltipContent className="bg-red-500">Delete semester</TooltipContent>
+                </Tooltip>
+            </CardTitle>
+            <Separator className="bg-emerald-500" />
+            <CardContent className="flex flex-col gap-y-4">
+                <div className="pt-4 grid grid-cols-10 gap-x-4">
+                    <p className="pl-1 col-span-3 font-semibold text-sm uppercase">Subject name</p>
+                    <p className="pl-1 col-span-3 font-semibold text-sm uppercase">Grade %</p>
+                    <p className="pl-1 col-span-3 font-semibold text-sm uppercase">Credits</p>
+                    <div className="col-span-1"></div>
                 </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>Predictive Analytics</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester6.predictiveAnalytics]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester6: {
-                                ...prev.semester6,
-                                predictiveAnalytics: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester6.predictiveAnalytics}%</p>
+
+                {subjects.fields.map((_, subjectIndex) => (
+                    <div key={subjectIndex} className="grid grid-cols-10 gap-x-4">
+                        <FormField
+                            control={form.control}
+                            name={`semesters.${semesterIndex}.subjects.${subjectIndex}.name`}
+                            render={({ field, fieldState }) => (
+                                <div className="flex flex-col col-span-3">
+                                    <Input
+                                        {...field}
+                                        className={cn("", fieldState.error && "border-red-500")}
+                                        placeholder="Subject name"
+                                    />
+                                    <p className="text-red-500">{fieldState.error?.message}</p>
+                                </div>
+                            )}
+                        />
+                        <FormField
+                            control={form.control}
+                            name={`semesters.${semesterIndex}.subjects.${subjectIndex}.grade`}
+                            render={({ field, fieldState }) => (
+                                <div className="flex flex-col col-span-3">
+                                    <Input
+                                        {...field}
+                                        onChange={(e) => {
+                                            // Inputs are always strings, so we need to convert them to numbers
+                                            field.onChange(Number(e.target.value))
+                                        }}
+                                        type="number"
+                                        className={cn("", fieldState.error && "border-red-500")}
+                                        placeholder="Grade"
+                                    />
+                                    <p className="text-red-500">{fieldState.error?.message}</p>
+                                </div>
+                            )}
+                        />
+                        <FormField
+                            control={form.control}
+                            name={`semesters.${semesterIndex}.subjects.${subjectIndex}.credits`}
+                            render={({ field, fieldState }) => (
+                                <div className="flex flex-col col-span-3">
+                                    <Input
+                                        {...field}
+                                        onChange={(e) => {
+                                            // Inputs are always strings, so we need to convert them to numbers
+                                            field.onChange(Number(e.target.value))
+                                        }}
+                                        type="number"
+                                        className={cn("", fieldState.error && "border-red-500")}
+                                        placeholder="Credits"
+                                    />
+                                    <p className="text-red-500">{fieldState.error?.message}</p>
+                                </div>
+                            )}
+                        />
+                        <Tooltip>
+                            <TooltipTrigger asChild>
+                                <Button
+                                    size="icon"
+                                    variant="destructive"
+                                    onClick={() => subjects.remove(subjectIndex)}
+                                >
+                                    <TrashIcon />
+                                </Button>
+                            </TooltipTrigger>
+                            <TooltipContent className="bg-red-500">Delete course</TooltipContent>
+                        </Tooltip>
+                    </div>
+                ))}
+
+                <div className="flex gap-4 mt-6">
+                    <div className="flex-grow">
+                        <Tooltip>
+                            <TooltipTrigger asChild>
+                                <Button
+
+                                    size="icon"
+                                    onClick={() => subjects.append({ name: "", grade: 0, credits: 0 })}
+                                >
+                                    <PlusIcon />
+                                </Button>
+                            </TooltipTrigger>
+                            <TooltipContent>Add course</TooltipContent>
+                        </Tooltip>
+                    </div>
+                    <div>
+                        <p className="text-lg font-semibold">Mark: {gpa}% ({markToLabel(gpa)})</p>
+                    </div>
                 </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>Project Development</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester6.projectDevelopment]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester6: {
-                                ...prev.semester6,
-                                projectDevelopment: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester6.projectDevelopment}%</p>
-                </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>Computer Science</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester6.computerScience]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester6: {
-                                ...prev.semester6,
-                                computerScience: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester6.computerScience}%</p>
-                </div>
-            </div>
-            <div className="grid w-full max-w-sm items-center gap-1.5">
-                <Label>UX Design</Label>
-                <div className="flex gap-2">
-                    <Slider value={[markState.semester6.uxDesign]} min={0} max={100} onValueChange={(v) => {
-                        setMarkState((prev) => ({
-                            ...prev,
-                            semester6: {
-                                ...prev.semester6,
-                                uxDesign: v[0]!
-                            }
-                        }))
-                    }} />
-                    <p>{markState.semester6.uxDesign}%</p>
-                </div>
-            </div>
-        </div>
+            </CardContent>
+        </Card>
     )
 }
 
 export default function GPACalculator() {
-    const { toast } = useToast()
     const [hasLoaded, setHasLoaded] = useState(false)
-    const [markState, setMarkState] = useState<MarkType>({
-        semester5: {
-            computerVision: 0,
-            researchProject: 0,
-            softwareEngineering: 0,
-            devops: 0,
-            serverside: 0,
-        },
-        semester6: {
-            legalIssues: 0,
-            predictiveAnalytics: 0,
-            projectDevelopment: 0,
-            computerScience: 0,
-            uxDesign: 0,
+
+    // Form setup
+    const form = useForm<FormType>({
+        resolver: zodResolver(FormSchema),
+        defaultValues: {
+            semesters: []
         }
     })
 
-    const saveMarks = (marks: MarkType) => {
-        localStorage.setItem("marks", JSON.stringify(marks))
-    }
+    // Extract semesters field array
+    const semesters = useFieldArray({
+        control: form.control,
+        name: "semesters"
+    })
 
-    const loadMarks = () => {
-        const marks = localStorage.getItem("marks")
-
-        if (marks) {
-            setMarkState(JSON.parse(marks) as MarkType)
+    useEffect(() => {
+        const semesters = localStorage.getItem("semesters")
+        if (semesters) {
+            form.setValue("semesters", JSON.parse(semesters) as FormType["semesters"])
         }
-
         setHasLoaded(true)
+    }, [form])
+
+    // Save to local storage on change
+    const formWatch = form.watch()
+    useEffect(() => {
+        if (hasLoaded) {
+            const values = form.getValues()
+            localStorage.setItem("semesters", JSON.stringify(values.semesters))
+        }
+    }, [hasLoaded, formWatch, form])
+
+    // Will be called on form submission
+    const onSubmit = (_data: FormType) => {
+        return
     }
-
-    useEffect(() => {
-        loadMarks()
-    }, [])
-
-    useEffect(() => {
-        if (hasLoaded)
-            saveMarks(markState)
-    }, [markState, hasLoaded])
 
     return (
-        <div className="w-full flex-grow flex flex-col items-center  pt-10">
-            <h1 className="text-3xl font-bold">GPA calculator</h1>
-            <div className="w-[80%] flex flex-col gap-10 mt-20">
-                <div className="grid grid-cols-1 lg:grid-cols-2 gap-10 lg:gap-2 place-items-center flex-grow">
-                    <div className="h-full w-full flex flex-col items-center justify-center">
-                        <p className="text-xl">Semester 5 Grades</p>
-                        <Semester5Inputs markState={markState} setMarkState={setMarkState} />
-                        <div className="mt-8 w-fit px-4 p-2 rounded-lg bg-secondary border-2 border-green-400 text-secondary-foreground">
-                            Total: {calculateS5Total(markState)}% ({getMarkLabel(calculateS5Total(markState))})
+        <TooltipProvider>
+            <Form {...form}>
+                <form onSubmit={form.handleSubmit(onSubmit)} className="items-center flex flex-col h-full">
+                    <div className="fixed bottom-10 right-10">
+                        <Tooltip>
+                            <TooltipTrigger asChild>
+                                <Button
+                                    className="h-12 aspect-square"
+                                    onClick={() => semesters.append({ name: "", subjects: [{ name: "", grade: 0, credits: 0 }] })}
+                                >
+                                    <PlusIcon />
+                                </Button>
+                            </TooltipTrigger>
+                            <TooltipContent>Add semester</TooltipContent>
+                        </Tooltip>
+                    </div>
+                    <div className="flex flex-col h-full gap-10">
+                        <div className="text-center flex flex-col gap-1">
+                            <h1 className="text-2xl font-bold">GPA Calculator</h1>
+                            <p className="text-md font-light">Calculate your GPA with this tool</p>
+                        </div>
+
+                        {semesters.fields.length === 0 && (
+                            <div className="flex flex-col flex-grow items-center justify-center text-center">
+                                <p className="text-3xl font-bold text-primary">You haven&apos;t added any semesters yet! </p>
+                                <p className="text-xl font-thin text-primary">Click the button in the corner to add a semester</p>
+                            </div>
+                        )}
+
+                        <div className="grid grid-cols-1 xl:grid-cols-2 gap-8 pb-32 lg:pb-10">
+                            {semesters.fields.map((semester, semesterIndex) => (
+                                <Semester
+                                    key={semester.id}
+                                    semesterIndex={semesterIndex}
+                                    semestersFieldArray={semesters}
+                                    form={form}
+                                />
+                            ))}
                         </div>
                     </div>
-                    <div className="h-full w-full flex flex-col items-center justify-center">
-                        <p className="text-xl">Semester 6 Grades</p>
-                        <Semester6Inputs markState={markState} setMarkState={setMarkState} />
-                        <div className="mt-8 w-fit px-4 p-2 rounded-lg bg-secondary border-2 border-green-400 text-secondary-foreground">
-                            Total: {calculateS6Total(markState)}% ({getMarkLabel(calculateS6Total(markState))})
-                        </div>
-                    </div>
-                </div>
-                <div className="col-span-2 w-full flex flex-col items-center justify-center size-fit">
-                    <div className="bg-secondary rounded-lg p-2 w-fit px-4 border-2 border-green-600">
-                        Combined Total: {calculateTotal(markState)}% ({getMarkLabel(calculateTotal(markState))})
-                    </div>
-                    <div className="gap-2 grid grid-cols-2 mt-2">
-                        <Button variant="default" onClick={() => { saveMarks(markState); toast({ title: "Saved!" }) }} className="w-full">Save</Button>
-                        <Button variant="destructive" onClick={() => setMarkState({
-                            semester5: {
-                                computerVision: 0,
-                                researchProject: 0,
-                                softwareEngineering: 0,
-                                devops: 0,
-                                serverside: 0,
-                            },
-                            semester6: {
-                                legalIssues: 0,
-                                predictiveAnalytics: 0,
-                                projectDevelopment: 0,
-                                computerScience: 0,
-                                uxDesign: 0,
-                            }
-                        })} className="w-full">Reset</Button>
-                    </div>
-                </div>
-            </div>
-        </div>
+                </form>
+            </Form>
+        </TooltipProvider>
     )
 }
-
